Make attribute lookup more explicit

+++ src/Rendering/DSCS/DataObjects/OpenGLDSCSMesh.cpp
@@ -25,7 +25,7 @@ namespace Rendering::DSCS::DataObjects
 
 		for (const auto& va : this->mesh.vertex_attributes)
 		{
-			this->vertex_attributes.emplace_back(OpenGLDSDSVertexAttribute{ va_map_DSCS_to_GL[(uint16_t)va.attribute_type], va.num_elements, dtype_map_DSCS_to_GL[va.data_type], va.always_20, va.vertex_struct_offset });
+			this->vertex_attributes.emplace_back(OpenGLDSDSVertexAttribute{ va_map_DSCS_to_GL[va.attribute_type], va.num_elements, dtype_map_DSCS_to_GL[va.data_type], va.always_20, va.vertex_struct_offset });
 		}
 	}
